# fix: out-of-memory crash (#7720) and significantly improve performance + feat: extracting fields to types