--- a/packages/plugins/other/visitor-plugin-common/src/base-documents-visitor.ts
+++ b/packages/plugins/other/visitor-plugin-common/src/base-documents-visitor.ts
@@ -30,6 +30,7 @@ function getRootType(operation: OperationTypeNode, schema: GraphQLSchema) {
 export interface ParsedDocumentsConfig extends ParsedTypesConfig {
   addTypename: boolean;
   preResolveTypes: boolean;
+  extractAllFieldsToTypes: boolean;
   globalNamespace: boolean;
   operationResultSuffix: string;
   dedupeOperationSuffix: boolean;
@@ -275,6 +276,11 @@ export class BaseDocumentsVisitor<
     );
     const operationType: string = pascalCase(node.operation);
     const operationTypeSuffix = this.getOperationSuffix(name, operationType);
+    const selectionSetObjects = selectionSet.transformSelectionSet(
+      this.convertName(name, {
+        suffix: operationTypeSuffix,
+      })
+    );
 
     const operationResult = new DeclarationBlock(this._declarationBlockConfig)
       .export()
@@ -284,7 +290,7 @@ export class BaseDocumentsVisitor<
           suffix: operationTypeSuffix + this._parsedConfig.operationResultSuffix,
         })
       )
-      .withContent(selectionSet.transformSelectionSet()).string;
+      .withContent(selectionSetObjects.mergedTypeString).string;
 
     const operationVariables = new DeclarationBlock({
       ...this._declarationBlockConfig,
@@ -299,6 +305,23 @@ export class BaseDocumentsVisitor<
       )
       .withBlock(visitedOperationVariables).string;
 
-    return [operationVariables, operationResult].filter(r => r).join('\n\n');
+    const dependentTypesContent = this._parsedConfig.extractAllFieldsToTypes
+      ? selectionSetObjects.dependentTypes.map(
+          i =>
+            new DeclarationBlock(this._declarationBlockConfig)
+              .export()
+              .asKind('type')
+              .withName(i.name)
+              .withContent(i.content).string
+        )
+      : [];
+
+    return [
+      ...(dependentTypesContent.length > 0 ? [dependentTypesContent.join('\n')] : []),
+      operationVariables,
+      operationResult,
+    ]
+      .filter(r => r)
+      .join('\n\n');
   }
 }