… dev-test/gatsby/schema.graphql
… packages/plugins/other/visitor-plugin-common/src/selection-set-to-object.ts
… packages/plugins/typescript/operations/tests/extract-all-types.spec.ts