@charpeni 2023-10-12
I'll spend more time reviewing this in detail, but just wanted to reflect on the fact that the results are, indeed, jaw-dropping! Big unions and intersections come with a heavy price tag!

This is amazing, thank you for contributing and for the well-thought-out description. It was interesting to read. 🙌 

I look forward to reviewing this in detail.